+++ Editor/Scripts/SubPackage/SubPackageConfigSchema.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+using UnityEngine.Serialization;
+
+namespace SubPackage
+{
+
+    [System.Serializable]
+    class SubPackageConfigSchema
+    {
+        public string dialogTitle;
+        public string dialogText;
+
+        public string cleanupRegex;
+        public SubPackageEntrySchema[] subPackages;
+    }
+
+}